+++ code/g3dengine/g3dsysinfo.c
@@ -38,7 +38,7 @@ void g3dsysinfo_processAnimatedImages ( G3DSYSINFO *info,
                                         float       sceneCurrentFrame,
                                         float       sceneEndFrame,
                                         float       sceneFramesPerSecond,
-                                        uint32_t    engine_flags ) {
+                                        uint64_t engine_flags ) {
     LIST *ltmpAnimatedImages = info->lanimatedImages;
 
     while ( ltmpAnimatedImages ) {
