+++ code/g3dmouse/gloss3d/createFacegroup.c
@@ -38,7 +38,7 @@ static uint32_t createFacegroup_init  ( G3DMOUSETOOL *mou,
                                         G3DSCENE *sce, 
                                         G3DCAMERA *cam,
                                         G3DURMANAGER *urm, 
-                                        uint32_t engine_flags );
+                                        uint64_t engine_flags );
 
 /******************************************************************************/
 G3DMOUSETOOLCREATEFACEGROUP *g3dmousetoolcreatefacegroup_new ( ) {
@@ -67,7 +67,7 @@ static uint32_t createFacegroup_init  ( G3DMOUSETOOL *mou,
                                         G3DSCENE *sce, 
                                         G3DCAMERA *cam,
                                         G3DURMANAGER *urm, 
-                                        uint32_t engine_flags ) {
+                                        uint64_t engine_flags ) {
     G3DOBJECT *obj = g3dscene_getLastSelected ( sce );
 
     if ( ( obj ) && ( obj->type & MESH ) ) {
